(feat) move base files and eterbase to new location
