+++ hummingbot/connector/exchange/eterbase/eterbase_auth.py
