+++ hummingbot/connector/exchange/eterbase/eterbase_order_book_tracker.py
@@ -19,13 +19,13 @@ from hummingbot.core.event.events import TradeType
 from hummingbot.logger import HummingbotLogger
 from hummingbot.core.data_type.order_book_tracker import OrderBookTracker, OrderBookTrackerDataSourceType
 from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
-from hummingbot.market.eterbase.eterbase_api_order_book_data_source import EterbaseAPIOrderBookDataSource
-from hummingbot.market.eterbase.eterbase_order_book_message import EterbaseOrderBookMessage
+from hummingbot.connector.exchange.eterbase.eterbase_api_order_book_data_source import EterbaseAPIOrderBookDataSource
+from hummingbot.connector.exchange.eterbase.eterbase_order_book_message import EterbaseOrderBookMessage
 from hummingbot.core.data_type.order_book_message import (
     OrderBookMessageType)
-from hummingbot.market.eterbase.eterbase_order_book import EterbaseOrderBook
-from hummingbot.market.eterbase.eterbase_active_order_tracker import EterbaseActiveOrderTracker
-import hummingbot.market.eterbase.eterbase_constants as constants
+from hummingbot.connector.exchange.eterbase.eterbase_order_book import EterbaseOrderBook
+from hummingbot.connector.exchange.eterbase.eterbase_active_order_tracker import EterbaseActiveOrderTracker
+import hummingbot.connector.exchange.eterbase.eterbase_constants as constants
 
 
 class EterbaseOrderBookTracker(OrderBookTracker):
