+++ hummingbot/connector/exchange/eterbase/eterbase_trading_rule.pxd
