+++ test/integration/test_eterbase_market.py
@@ -38,8 +38,8 @@ from hummingbot.core.utils.async_utils import (
     safe_ensure_future,
     safe_gather,
 )
-from hummingbot.market.eterbase.eterbase_market import EterbaseMarket
-from hummingbot.market.eterbase.eterbase_utils import convert_to_exchange_trading_pair
+from hummingbot.connector.exchange.eterbase.eterbase_market import EterbaseMarket
+from hummingbot.connector.exchange.eterbase.eterbase_utils import convert_to_exchange_trading_pair
 from hummingbot.market.market_base import OrderType
 from hummingbot.connector.markets_recorder import MarketsRecorder
 from hummingbot.model.market_state import MarketState
