+++ test/integration/test_eterbase_user_stream_tracker.py
@@ -20,9 +20,9 @@ from hummingbot.core.utils.async_utils import (
     safe_ensure_future,
     safe_gather,
 )
-from hummingbot.market.eterbase.eterbase_order_book_message import EterbaseOrderBookMessage
-from hummingbot.market.eterbase.eterbase_market import EterbaseMarket, EterbaseAuth
-from hummingbot.market.eterbase.eterbase_user_stream_tracker import EterbaseUserStreamTracker
+from hummingbot.connector.exchange.eterbase.eterbase_order_book_message import EterbaseOrderBookMessage
+from hummingbot.connector.exchange.eterbase.eterbase_market import EterbaseMarket, EterbaseAuth
+from hummingbot.connector.exchange.eterbase.eterbase_user_stream_tracker import EterbaseUserStreamTracker
 from hummingbot.market.market_base import OrderType
